requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.4.2
aiohttp>=3.8.6
tenacity>=8.2.3
google-generativeai>=0.3.0